email_backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
python-jose==3.3.0
gunicorn==20.1.0
requests==2.26.0
werkzeug==2.0.1 